Merged PR 17164: Support Running on AZP Machines
AZP is rolling out test signing support on the `windows-2022` images, and they asked me to validate things. This PR is the result of trying out both the functional tests and spinxsk. There are still a few issues to work on, but generally the test signing part is working! For now, it keeps the new stuff disabled, but hopefully we can enable it soon.