Merged PR 17259: Add OutDir to build_headers binaries so they don't litter toplevel xdp directory
I created the vcxproj files for build_headers without OutDir properties since we don't need
the binaries, but instead of keeping the binary from being created the result is that the
binaries are placed in the current directory (i.e. the toplevel xdp directory), which litters
the repo.

So, I've added OutDir properties to send the files to artifacts\bin.