Merged PR 17258: Split readme.md into development.md and usage.md
Currently readme.md is included in the devkit, but has info that doesn't make sense
in the context of a devkit. The documentation can be pretty cleanly split into a document
on "how to use xdp" and one on "how to change xdp", so I've done that, and included
only the "how to use" doc in the devkit.

I've also created a "readme" directory to contain doc files; I foresee that we'll eventually
add an md file describing the overall code architecture and so on.