Merged PR 17339: XDP: add WPP tracing to spinxsk
Add skeleton for WPP event tracing to spinxsk.

It turns out that regular EXE and DLL builds do not have built-in support for WPP configuration via .vcxproj files, so we have to invoke `tracewpp.exe` manually. This requires a bit of nasty macro guessing where the WPP configuration resides, but we'll gloss over that.