Merged PR 17401: Fix reboot flag in prepare-machine.ps1
Fix reboot flag in prepare-machine.ps1

When preparing a test machine by invoking the script with
`-ForTest:$true -NoReboot:$false`, output would indicate that
testsigning was set and that a reboot was needed, but no reboot
occurred.

About powershell scopes from MSDN:
"An item that you created within a scope can be changed only in the
scope in which it was created, unless you explicitly specify a
different scope."
"A child scope does not inherit the variables, aliases, and functions
from the parent scope. Unless an item is private, the child scope can
view the items in the parent scope. And, it can change the items by
explicitly specifying the parent scope, but the items are not part of
the child scope."

So setting "$Reboot = $true" initialized a local scope var instead of
writing to the script scope var.