Merged PR 17468: Fix crash on fndis startup
This crash occurs on fndis startup when verifier is enabled with fault injection. DriverEntry
calls NdisPollCpuStart, which successfully allocates PollCpus, but then one of the
IoAllocateWorkItem calls fails. STATUS_NO_MEMORY is returned to DriverEntry, so
DriverUnload->NdisPollCpuStop is called, and since PollCpus is non-NULL, KeInsertQueueDpc
is called for each NDIS_POLL_CPU's DPC. Since we didn't finish initializing, KeInitializeDpc wasn't
called for at least one of those DPCs. Surprisingly, KeInsertQueueDpc seems to only validate DPCs
for things like "the callback routine is not NULL" on x86 architecture, so instead of crashing in
an obvious way, we crash later when the kernel is flushing DPCs and tries to call a NULL function
pointer.