Merged PR 17507: Replace all usages of Write-Debug with Write-Host in powershell scripts
Write-Debug messages only appear when you run a script with -Debug. And in
powershell 5, when you run a script with -Debug, Write-Debug pauses the script's
execution until the user enters "Y". This bug has been fixed in powershell 7, but there
are apparently no plans to fix it in powershell 5, and we should not require powershell 7.

Some people still want to muck around with verbosity settings in powershell, so instead of
Write-Host, Write-Verbose has been used. So you still have to pass an extra flag to most
scripts most of the time.