Merged PR 17643: Add binding set abstraction to xdpif.h
Add binding set abstraction to xdpif.h

To prep for the introduction of offloads scoped to the NIC interface
level (ex: RSS), add binding set as an abstraction to xdpif.h.

There is a lot of confusing/inconsistent terminology near the xdpif
boundary. This PR tries to clear up some, but not all issues.