Merged PR 17678: Update development.md manual spinxsk verifier command to match CI
For spinxsk in CI we run verifier only on xdp.sys. Updates development.md to match.
I've wasted a lot of time in the past due to confusion about xdpmp load failures that were
caused by verifier fault injection, and thought I was configuring things the same as the CI,
so it's important to keep the two in sync.