Merged PR 17763: XDP: fix devkit symbols and rev version
Mellanox is hitting a linker error because our devkit is missing symbols for the NMR static library. Add the symbols to the devkit and rev the version.

I'm not sure why this wasn't an issue before - it's possible the OS build system uses an older symbols format that gets included along with static libraries.