Merged PR 17785: Rename Binding to Interface in core bind.c
Rename Binding to Interface in core bind.c

After agreeing on "interface" and "interface set" primitives for the
XDPIF API, rename XDP_INTERFACE related variables and functions to
match. This change does not update the binding API that the core bind
module exposes to the rest of XDP core. Changes include:

- Rename XdpBindingLock to XdpInterfaceSetsLock
- Rename "XDP_INTERFACE Binding" variables to "XDP_INTERFACE Interface"
- Rename static functions referring to "Binding" to "Interface"
- Rename XdpIfpReferenceInterface/XdpIfpDereferenceInterface to
  XdpIfpReferenceProvider/XdpIfpDereferenceProvider to increase clarity
  and avoid a fn name collision after the above bullet