Merged PR 17908: XDP: fix EC race condition on cleanup
I spent an hour squinting at the EC code and a crash dump, and finally managed to spot the bug: there is a race condition where the `XdpEcPoll` routine can re-arm the EC, check the cleanup event, and set the completion event before `XdpEcCleanup` sends the gratuitous notification that guarantees the polling state machine will run. If this race happens, the gratuitous notification will queue a DPC, which is no good since the cleanup event is already set.

Alternatively we could try to fix this by disarming the queue within `XdpEcCleanup` but that fix is even more complicated.