Merged PR 18339: Ref count RSS offload setting
Ref count RSS offload setting

When an XDP program or an XSK implicitly reference a queue (programs
and sockets must be associated with a queue ID) the given RSS
configuration should not change for the lifetime of the progam/XSK.
So the RSS configuration will need to be ref counted.

This PR refactors the RSS offload code to use a new, ref counted RSS
settings type. Plumbing references from programs/XSKs will come in a
subsequent PR.

Also consolidate the logic of removing the lower edge RSS settings
and plumbing down the upper edge settings.