Merged PR 18375: Active XDP objects reference RSS settings
Active XDP objects reference RSS settings

When an XDP program or an XSK implicitly reference a queue (programs
and sockets must be associated with a queue ID) the given RSS
configuration should not change for the lifetime of the progam/XSK.
So the RSS configuration will need to be ref counted.

This PR adds the plumbing for these references. Associated test cases
are added/enabled.