Merged PR 18616: XDP: Add WPP tracing to functional tests
There's a bit of log spew introduced during the RSS work, and it outputs to the VS test console. I imagine this is temporary debugging and was intended to be removed, but this exposes the fact that we are missing WPP tracing in functional tests.

Rather than continue the pattern of adding tracing during test development, then ripping it out once the tests stabilize, convert the few existing traces to WPP and collect automatically during test runs.