Merged PR 18651: XDP: add ETW events for RX and TX latency measurements
Add minimum viable ETW events for RX and TX latency measurements for MsQuic over generic XDP.

On the TX path, MsQuic can correlate each of its socket handles to a kernel XSK and its underlying generic TX queue. We emit events for each `XskNotifySocket`, each XSK frame that gets submitted to the generic TX queue, each XDP frame that gets added to an NBL chain, and each NBL chain that gets submitted to NDIS.

Individual frames can be tracked from the application layer down to NDIS: Each `XskTxIndex` is identical to the user mode XSK TX ring producer index of a frame, which is then mapped by XSK into a separate XDP ring for kernel TX. For generic XDP, each of these kernel XDP frames is appended to a chain of NBLs, and each NBL chain is passed to NDIS.

```
[0]1348.0C04::?2022?-?01?-?07 11:36:05.546 [Microsoft-XDP][ xsk][0xFFFFFD8F9D2E2D40] create socket ProcessId=0x1348
[2]0004.0DB0::?2022?-?01?-?07 11:36:05.547 [Microsoft-XDP][ xsk][0xFFFFFD8F9D2E2D40] bind TX XdpInterfaceQueue=0xFFFFFD8F9D1EEEB0
[0]1348.0C04::?2022?-?01?-?07 11:36:05.548 [Microsoft-XDP][ xsk][0xFFFFFD8F9D2E2D40] notify poke Flags=0x2
[0]1348.0C04::?2022?-?01?-?07 11:36:05.548 [Microsoft-XDP][ xsk][0xFFFFFD8F9D2E2D40] enqueue TX XskTxIndex=0x0 XdpTxIndex=0x0
[0]1348.0C04::?2022?-?01?-?07 11:36:05.548 [Microsoft-XDP][ xsk][0xFFFFFD8F9D2E2D40] enqueue TX XskTxIndex=0x1 XdpTxIndex=0x1
[0]1348.0C04::?2022?-?01?-?07 11:36:05.548 [Microsoft-XDP][gxtq][0xFFFFFD8F9D1EEEB0] enqueue TX XdpTxIndex=0x0 NblBatchIndex=0x0
[0]1348.0C04::?2022?-?01?-?07 11:36:05.548 [Microsoft-XDP][gxtq][0xFFFFFD8F9D1EEEB0] enqueue TX XdpTxIndex=0x1 NblBatchIndex=0x0
[0]1348.0C04::?2022?-?01?-?07 11:36:05.548 [Microsoft-XDP][gxtq][0xFFFFFD8F9D1EEEB0] post TX batch NblBatchIndex=0x0
[0]1348.0C04::?2022?-?01?-?07 11:36:05.548 [Microsoft-XDP][gxtq][0xFFFFFD8F9D1EEEB0] post TX batch complete NblBatchIndex=0x0
[0]1348.0C04::?2022?-?01?-?07 11:36:05.548 [Microsoft-XDP][ xsk][0xFFFFFD8F9D2E2D40] notify poke complete Flags=0x2
[0]1348.0C04::?2022?-?01?-?07 11:36:05.548 [Microsoft-XDP][ xsk][0xFFFFFD8F9D2E2D40] close socket
[0]1348.0C04::?2022?-?01?-?07 11:36:05.549 [Microsoft-XDP][ xsk][0xFFFFFD8F9D2E2D40] close socket complete
```

On the RX path, all processing occurs inline, and therefore less opportunities for latency are possible. We emit a pair of events for the NDIS entry/exit points (i.e. when NDIS hands us a batch of NBLs to inspect, and when we later return control back to NDIS) and one event after a batch of frames has been passed to the XSK ring.

```
[0]1348.0C04::?2022?-?01?-?07 11:37:17.549 [Microsoft-XDP][ xsk][0xFFFFFD8F9D11CD40] create socket ProcessId=0x1348
[0]1348.0C04::?2022?-?01?-?07 11:37:17.550 [Microsoft-XDP][gxrf][0xFFFFFD8F9D102E50] inspect RX batch
[0]1348.0C04::?2022?-?01?-?07 11:37:17.550 [Microsoft-XDP][ xsk][0xFFFFFD8F9D11CD40] post RX batch XskRxIndex=0x0 BatchSize=0x1
[0]1348.0C04::?2022?-?01?-?07 11:37:17.550 [Microsoft-XDP][gxrf][0xFFFFFD8F9D102E50] inspect RX batch com...