Merged PR 19848: XDP: convert LWF to batched receive inspection
To prepare for removal of the single (i.e. unbatched) receive API, convert the LWF to use the batched receive API. Due to the complexity of the NDIS data path contract, this unfortunately further complicates the LWF data path, but the ultimate removal of the single receive API should be a net simplification for the project.