Merged PR 19976: Remove XskNotifySocket IRP path
Remove XskNotifySocket IRP path

Under the assumption that the IO manager always attempts fast IO if
there is a handler registered, the IOCTL_XSK_NOTIFY IRP path is dead.
Additionally the code was found to be buggy. Remove the path entirely.

The bugs that were causing XskNotifySocket to fail when the IRP (not
FastIo) IOCTL path is forced:

1) XskIrpDeviceIoControl was incorrectly passing the FILE_OBJECT to
   XskNotify instead of FILE_OBJECT->FsContext (XSK)

2) XskIrpDeviceIoControl calls XskNotify with
   Irp->AssociatedIrp.SystemBuffer, which is a kernel-mapped address.
   XskNotifyValidateParams probes the InputBuffer address if the
   previous mode wasn't kernel (the previous mode is user mode, even
   for buffered IOCTLs) and so we fail.