Merged PR 20512: XDP: set NeedPoke flag on TX ring when blocked by full TX completion ring
Today it is possible for an application with a non-empty TX ring and full TX completion ring to fail to poke the driver after making space in its TX completion ring.

It seems the original design (copied from Linux, presumably) was that XskNotifySocket should return a failure result in this case, but since our XDP supports asynchronous completion of poke requests, we can't set the output flags on a per-IO basis.

Instead, we obligate applications to invoke XskNotifySocket after dequeuing elements from the TX completion ring when elements are already on the TX ring, and rely on the updated need poke flag to elide unnecessary system calls.

Update tests to follow this pattern.