Merged PR 20616: XDP: revert regression: don't parse RX fragments unless necessary
Looking at the LoLa xperfs, there weren't any perf surprises except that we're taking the slower fragment-handling path. This is unexpected since we'd expect the NBLs indicated by mellanox to be contiguous.

It turns out a PR removed the early return when all headers are parsed in the first fragment, so restore that optimization.

!18957