Merged PR 20621: XDP: remove 32-bit process checks
The only surprise in the xperfs from QUIC tests is that 32-bit process checks show up on the hot path in XskNotifySocket. The contribution to CPU usage is miniscule, but still, there's no reason to check for 32-bit processes after the handle is opened since we don't support 32-bit in the first place. The driver shouldn't do anything harmful, but obviously apps will get undefined behavior.