Merged PR 20647: XDP: remove RSS config reference on RX/TX queues
This has never been removed, but the idea that it might be convenient for an XSK to be associated with an RSS config handle became mangled into the idea that any XSK should prevent the RSS configuration changing.

The convenient handle association has not been implemented and is pretty low-value. Remove the mangled reference scheme, but leave behind the offload handle within RX/TX queues because those will soon be useful.