Merged PR 20649: XDP: work around NDIS registry reset powershell glitches
I've seen two PRs fail to run `Reset-NetAdapterAdvancedProperty -DisplayName *` cleanly, and my hypothesis is something goes wonky between resetting the registry keys and restarting the miniport.

Testing in a local PS console, I was able to repro the error with a small `for` loop, and by separating the restart from the reset I was able to get the loop to run clean.

```powershell
for ($i = 0; $i -lt 100; $i++) {
    Reset-NetAdapterAdvancedProperty -Name XDPFNMP -DisplayName * -NoRestart
    Restart-NetAdapter XDPFNMP
    Write-Progress -Activity RestartTest -PercentComplete $i
}
```