Merged PR 20697: XDP: simplify test setup steps
One of our functional tests failed in the CI with no obvious explanation, so I tried to repro manually. This turned out to be unnecessarily complicated, so streamline the process.

Also fix a longstanding bug that ndis.sys was not verified in functional tests. I've already updated the 1ES pools.