Merged PR 20811: XDP: refactor LWF delay detach
Simplify the the LWF delayed detach logic: now, releasing the final reference causes a delayed detach timer to start, and acquiring a new reference causes the timer to be cancelled. Tearing down the interface cancels the timer.

This is conceptually much simpler than a worker thread that loops indefinitely while waiting for a continuously updated delayed detach interval to expire, while also monitoring for interface removal.

If we get around to improving the delayed detach regkey updates, this timer-based approach will be much simpler to use.