Merged PR 20808: XDP: ensure default delay detach cleanup takes effect
Follow up to !20754

The LWF delay detach parameter isn't reliably updated until all the data path references go away and the delayed detach thread exits, which can take a long time. The delay deatch test already restarts the miniport stack to work around this constraint, except in the final cleanup case. Add a restart there, too.

Another option is to improve the implementation of the delayed detach so the regkey takes effect immediately, but this is admittedly an esoteric feature and probably not worth the time.