Merged PR 20833: XDP: add port bitmap match conditions
To provide a relatively straightforward optimization to the QUIC client code path, add a few match conditions based on the destination UDP port, optionally combined with the local IP address. This means that binding/unbinding a local UDP port can be implemented via a handful of instructions and shared memory. In a local microbenchmark, it's possible to update over 300 billion ports per second, so this should be a worthy optimization.