Fix a race condition in xdp receive path in low resources scenario (#825)
## Background

In low resource scenario, XDP must return the NBL chain unchanged to NDIS.

To do so, in `XdpGenericReceivePostInspectNbs`, it flushes NetBuffers that need to be passed to the next driver in `XdpGenericReceiveLowResources`, so it can re-assemble the NBL chain in order (if waiting as in the non-low resource path, in-order reassembly would become complex, since NBLs would be split between a "drop" and "pass" lists).

To optimize this logic, the code tries to indicate batches of NBLs when possible, by indicating a sequence of packets only when a packet needs to be dropped. The batch of NBLs is easy to put back in order: first all the indicated "PASS" NBLs, then the DROP NBL.
The following sequence of action should result in two indications: one when a packet is dropped, one at the end of the sequence.

> PASS -> PASS -> PASS -> DROP [indicate] -> PASS -> PASS [indicate]

`XdpGenericReceiveLowResources` must releas and re-acquire the execution context lock when indicating packets:
another thread can start queuing packets for inspection at that point, so the state of the ring buffer must be consistent when releasing the lock.

To ensure that, in low resource mode, `XdpGenericReceivePreInspectNbs` add a single NBL at a time in XDP ring buffer. This packet can be grouped with NBL that are skipping the inspection and are not added to the ring buffer (loopback, failure to map the VA...)

> SKIP -> ADD_TO_RING [inspect] -> ADD_TO_RING [inspect] || SKIP -> SKIP -> ADD_TO_RING [inspect]

The reasoning likely being that this way, when indicating NBLs, the NBL from the ring would have already been retrieved and the ring be in a clean state for another thread. 

This is wrong on several scenarios.

## What is wrong

### First issue: A flush is done only after processing the full chain

For this sequence:

> ADD_TO_RING+PASS -> SKIP+DROP -> ADD_TO_RING+PASS

(NBL skipped because of fault injection or mapping failure)

- After the first inspection, the pass list contains the first NBL.
- The second pre-inspection will skip one NBL and add one to the ring
- The second post-inspection sees the DROP for the skipped NBL first, with an NBL present in the PASS list
    - it indicates the first PASS NBL, with the second PASS NBL still in the ring => ERROR

### Second issue: Loopback packets default to PASS

For this sequence:

> SKIP+PASS -> SKIP+DROP -> ADD_TO_RING+PASS

(first NBL skipped because of loopback, second because of fault injection or mapping failure)

- The SKIP+DROP NBL will cause the SKIP+PASS NBL to be indicated with an NBL still on the ring ==> ERROR

## Solution

No longer group NBL skipping inspection with NBL skipping inspection: only consider a single NBL per "pre / inspect / post" loop. This preserve the batching of the indication up the stack and remove all risk of invalid ring buffer state.